Update module formats and dependencies; Remove uuid package
This commit updates the module formats in tsup.config.ts to include "cjs" and "iife". Additionally, the "uuid" package and its types have been removed as dependencies in package.json and package-lock.json, and the nanoid package is now being used in model.ts. The package version has also been updated to 0.0.39.